+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unrealeased]
+### Added
+- Action now detects and fails on misplaced files
+### Changed
+- Improved output log for clarity
 
 ## [1.1.0]
 ### Added